@@ -1,22 +1,24 @@
 {
   "name": "project-sw",
   "version": "0.0.0",
+  "type": "module",
   "scripts": {
     "dev": "vite",
     "build": "vite build",
     "preview": "vite preview"
   },
   "dependencies": {
-    "@heroicons/vue": "^1.0.6",
-    "axios": "^0.27.2",
-    "vue": "^3.2.33",
-    "vue-router": "^4.0.15"
+    "@heroicons/vue": "^2.1.5",
+    "axios": "^1.7.4",
+    "vue": "^3.4.37",
+    "vue-router": "^4.4.3"
   },
   "devDependencies": {
-    "@vitejs/plugin-vue": "^2.3.2",
-    "autoprefixer": "^10.4.7",
-    "postcss": "^8.4.13",
-    "tailwindcss": "^3.0.24",
-    "vite": "^2.9.8"
+    "@iconify/vue": "^4.1.2",
+    "@vitejs/plugin-vue": "^5.1.2",
+    "autoprefixer": "^10.4.20",
+    "postcss": "^8.4.41",
+    "tailwindcss": "^3.4.10",
+    "vite": "^5.4.0"
   }
 }
