@@ -1,9 +1,12 @@
-module.exports = {
-  content: ["./index.html", "./src/**/*.{vue,js,ts,jsx,tsx}"],
+/** @type {import('tailwindcss').Config} */
+export default {
+  content: [
+    "./index.html",
+    "./src/**/*.{js,ts,vue}",
+  ],
   theme: {
-    extend: {
-      
-    },
+    extend: {},
   },
   plugins: [],
-};
+}
+
