@@ -18,12 +18,14 @@ import Navbar from "./components/Navbar.vue";
     url("/font/starjedi-webfont.woff") format("woff"),
     url("/font/starjedi.ttf") format("ttf");
 }
+
 @font-face {
   font-family: "Star Jhol";
   src: url("/font/starjhol-webfont.woff2") format("woff2"),
     url("/font/starjhol-webfont.woff") format("woff"),
     url("/font/starjhol.ttf") format("ttf");
 }
+
 #app {
   font-family: "Star Jedi";
   -webkit-font-smoothing: antialiased;
@@ -32,8 +34,9 @@ import Navbar from "./components/Navbar.vue";
   color: #ffe81f;
   letter-spacing: 3px;
 }
+
 body {
-  background: url("../public/background.jpg") no-repeat center center fixed;
+  background: url("../background.jpg") no-repeat center center fixed;
   background-size: cover;
 }
 
