@@ -2,29 +2,16 @@
 <template>
   <div class="relative pt-16 md:pt-24 pb-32 overflow-hidden">
     <div class="relative">
-      <div
-        class="lg:mx-auto lg:max-w-7xl lg:px-8 lg:grid lg:grid-cols-2 lg:grid-flow-col-dense lg:gap-24"
-      >
-        <div
-          class="px-4 max-w-xl mx-auto sm:px-6 lg:py-16 lg:max-w-none lg:mx-0 lg:px-0"
-        >
+      <div class="lg:mx-auto lg:max-w-7xl lg:px-8 lg:grid lg:grid-cols-2 lg:grid-flow-col-dense lg:gap-24">
+        <div class="px-4 max-w-xl mx-auto sm:px-6 lg:py-16 lg:max-w-none lg:mx-0 lg:px-0">
           <div>
             <div class="flex justify-center">
-              <span
-                class="h-12 w-12 rounded-md flex items-center justify-center bg-red-700 p-1"
-              >
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  x="0px"
-                  y="0px"
-                  width="50"
-                  height="50"
-                  class="fill-current h-full w-full"
-                  viewBox="0 0 50 50"
-                >
+              <span class="h-12 w-12 rounded-md flex items-center justify-center bg-red-700 p-1">
+                <svg xmlns="http://www.w3.org/2000/svg" x="0px" y="0px" width="50" height="50"
+                  class="fill-current h-full w-full" viewBox="0 0 50 50">
                   <path
-                    d="M 24 2 C 22.90625 2 22 2.90625 22 4 L 22 4.046875 C 16.664063 4.492188 12.289063 8.46875 11.242188 13.625 C 11.171875 13.902344 11.125 14.203125 11.089844 14.527344 C 11.074219 14.566406 11.023438 14.691406 11.003906 14.917969 C 10.9375 15.726563 10.734375 16.8125 10.484375 17.90625 C 10.40625 18.242188 10.328125 18.578125 10.246094 18.90625 L 10.246094 18.914063 C 10.238281 18.941406 10.226563 18.980469 10.214844 19.035156 C 9.929688 20.148438 9.632813 21.191406 9.40625 21.949219 C 9 22.964844 8.554688 24.039063 8.109375 25.109375 C 7.128906 27.457031 5.871094 30.128906 4.851563 32.589844 C 3.828125 35.042969 3.003906 37.21875 3 38.988281 C 2.996094 39.386719 3.132813 40.554688 4.328125 41.769531 C 5.097656 42.550781 6.273438 43.140625 7.882813 43.71875 C 9.355469 44.253906 11.261719 44.714844 13.371094 45.097656 C 13.652344 45.617188 14.042969 46.070313 14.554688 46.378906 C 15.289063 46.820313 16.1875 47.121094 17.234375 47.355469 C 19.332031 47.820313 22.03125 48 25 48 C 27.964844 48 30.667969 47.832031 32.765625 47.371094 C 33.8125 47.136719 34.714844 46.839844 35.453125 46.394531 C 35.964844 46.082031 36.355469 45.621094 36.636719 45.097656 C 38.742188 44.714844 40.648438 44.253906 42.121094 43.71875 C 43.722656 43.140625 44.902344 42.550781 45.671875 41.769531 C 46.867188 40.554688 47.003906 39.386719 47 38.988281 C 46.996094 37.21875 46.171875 35.046875 45.152344 32.589844 C 44.128906 30.128906 42.871094 27.457031 41.894531 25.109375 C 41.445313 24.035156 41 22.957031 40.589844 21.9375 C 40.574219 21.886719 40.5625 21.839844 40.546875 21.796875 L 40.546875 21.785156 C 40.550781 21.808594 40.546875 21.792969 40.535156 21.757813 L 40.535156 21.753906 C 40.5 21.625 40.457031 21.480469 40.410156 21.328125 C 40.398438 21.285156 40.386719 21.242188 40.375 21.191406 L 40.371094 21.1875 C 40.332031 21.054688 40.292969 20.917969 40.253906 20.78125 L 40.253906 20.777344 C 40.242188 20.734375 40.230469 20.699219 40.222656 20.667969 L 40.222656 20.664063 L 40.21875 20.660156 C 40.171875 20.488281 40.125 20.308594 40.074219 20.121094 L 40.070313 20.109375 C 40.058594 20.0625 40.046875 20.023438 40.039063 19.992188 L 40.035156 19.988281 L 40.035156 19.984375 C 39.996094 19.832031 39.953125 19.675781 39.910156 19.515625 C 39.894531 19.457031 39.882813 19.40625 39.875 19.363281 L 39.871094 19.355469 L 39.871094 19.347656 C 39.816406 19.152344 39.765625 18.949219 39.71875 18.75 L 39.707031 18.714844 L 39.707031 18.707031 C 39.660156 18.523438 39.617188 18.332031 39.570313 18.140625 L 39.566406 18.136719 C 39.554688 18.078125 39.542969 18.023438 39.53125 17.976563 L 39.53125 17.964844 C 39.273438 16.851563 39.066406 15.742188 38.996094 14.917969 C 38.976563 14.6875 38.925781 14.5625 38.910156 14.527344 C 38.875 14.203125 38.828125 13.898438 38.757813 13.625 C 37.710938 8.464844 33.335938 4.492188 28 4.046875 L 28 4 C 28 2.90625 27.09375 2 26 2 Z M 24 4 L 26 4 L 26 18 L 24 18 Z M 22 6.054688 L 22 17.59375 C 21.085938 17.3125 19.738281 17.003906 18 17 L 17 17 C 15.453125 17 13.832031 17.449219 12.4375 18.34375 C 12.695313 17.203125 12.914063 16.066406 12.996094 15.082031 C 12.992188 15.125 12.984375 15.164063 13.027344 15 L 13.046875 14.921875 L 13.058594 14.84375 C 13.089844 14.535156 13.132813 14.285156 13.1875 14.082031 L 13.191406 14.058594 L 13.199219 14.03125 C 14.054688 9.789063 17.621094 6.496094 22 6.054688 Z M 28 6.054688 C 32.378906 6.496094 35.945313 9.789063 36.800781 14.03125 L 36.808594 14.058594 L 36.8125 14.082031 C 36.867188 14.285156 36.910156 14.539063 36.941406 14.84375 L 36.953125 14.921875 L 36.972656 15 C 37.015625 15.164063 37.007813 15.125 37.003906 15.082031 C 37.085938 16.066406 37.304688 17.207031 37.566406 18.347656 C 36.167969 17.449219 34.546875 17 33 17 L 32 17 C 30.261719 17.003906 28.914063 17.3125 28 17.59375 Z M 19.953125 7.996094 C 19.863281 8 19.773438 8.019531 19.6875 8.046875 C 19.6875 8.046875 16.351563 9.136719 15.0625 12.65625 C 14.871094 13.171875 15.136719 13.75 15.65625 13.9375 C 16.171875 14.128906 16.75 13.863281 16.9375 13.34375 C 17.898438 10.738281 20.3125 9.953125 20.3125 9.953125 C 20.777344 9.808594 21.070313 9.347656 21.007813 8.867188 C 20.945313 8.382813 20.542969 8.015625 20.054688 7.996094 C 20.023438 7.996094 19.988281 7.996094 19.953125 7.996094 Z M 17 19 L 18 19 C 20.644531 19 22.574219 19.90625 22.574219 19.90625 L 22.777344 20 L 27.222656 20 L 27.425781 19.90625 C 27.425781 19.90625 29.355469 19 32 19 L 33 19 C 35.226563 19 37.667969 20.167969 38.726563 22.664063 C 39.144531 23.707031 39.597656 24.796875 40.046875 25.878906 C 41.050781 28.292969 42.308594 30.960938 43.304688 33.359375 C 44.300781 35.753906 45 37.945313 45 39 L 45 39.019531 C 45 38.933594 45.058594 39.542969 44.25 40.367188 L 44.246094 40.367188 C 43.90625 40.710938 42.898438 41.3125 41.4375 41.839844 C 40.15625 42.304688 38.503906 42.738281 36.59375 43.097656 C 36.445313 42.878906 36.226563 42.5 35.949219 41.675781 C 35.808594 41.25 35.402344 40.96875 34.953125 40.988281 C 34.632813 41.003906 34.34375 41.164063 34.164063 41.429688 C 33.988281 41.691406 33.945313 42.023438 34.050781 42.324219 C 34.367188 43.242188 34.683594 43.804688 34.925781 44.183594 C 34.871094 44.320313 34.757813 44.472656 34.414063 44.683594 C 33.976563 44.949219 33.253906 45.214844 32.335938 45.417969 C 30.496094 45.820313 27.886719 46 25 46 C 22.117188 46 19.507813 45.8125 17.671875 45.402344 C 16.75 45.199219 16.027344 44.929688 15.589844 44.667969 C 15.339844 44.515625 15.214844 44.394531 15.132813 44.289063 C 15.667969 43.742188 16.480469 42.761719 16.953125 41.308594 C 17.054688 41 17.003906 40.664063 16.816406 40.402344 C 16.628906 40.140625 16.324219 39.984375 16 39.984375 C 15.5625 39.988281 15.175781 40.273438 15.046875 40.691406 C 14.597656 42.09375 13.742188 42.902344 13.496094 43.113281 C 11.546875 42.753906 9.863281 42.308594 8.5625 41.839844 C 7.101563 41.3125 6.089844 40.710938 5.75 40.367188 C 4.941406 39.542969 5 38.933594 5 39.019531 L 5 39 C 5 37.945313 5.699219 35.753906 6.695313 33.359375 C 7.691406 30.960938 8.949219 28.292969 9.953125 25.878906 C 10.410156 24.792969 10.863281 23.695313 11.28125 22.644531 C 12.347656 20.164063 14.777344 19 17 19 Z M 18.359375 20 C 16.519531 20 15.082031 20.929688 14.222656 22.121094 C 13.402344 23.257813 13.042969 24.605469 13.011719 25.828125 C 13.011719 25.832031 13.011719 25.832031 13.011719 25.832031 C 13.003906 25.886719 13 25.945313 13 26 C 13 27.332031 13.53125 29.25 14.039063 31.027344 C 14.546875 32.800781 15.050781 34.316406 15.050781 34.316406 C 15.226563 34.839844 15.792969 35.125 16.316406 34.949219 C 16.839844 34.773438 17.125 34.207031 16.949219 33.683594 C 16.949219 33.683594 16.453125 32.199219 15.960938 30.472656 C 15.792969 29.890625 15.652344 29.285156 15.511719 28.6875 C 15.597656 28.710938 15.6875 28.75 15.769531 28.769531 C 16.851563 29.023438 17.785156 29.003906 17.929688 29 C 17.953125 29 17.976563 29 18 29 C 18 29 18.03125 29 18.03125 29 L 22 29 L 22 30.679688 L 17.1875 37.417969 C 16.96875 37.722656 16.9375 38.125 17.109375 38.457031 C 17.28125 38.792969 17.625 39 18 39 L 32 39 C 32.375 39 32.71875 38.792969 32.890625 38.457031 C 33.0625 38.125 33.03125 37.722656 32.8125 37.417969 L 28 30.679688 L 28 29 L 32 29 C 32 29 32.027344 28.996094 32.03125 29 C 32.125 29 33.117188 29.03125 34.230469 28.769531 C 34.3125 28.75 34.402344 28.710938 34.488281 28.6875 C 34.347656 29.285156 34.207031 29.890625 34.039063 30.472656 C 33.546875 32.199219 33.050781 33.683594 33.050781 33.683594 C 32.875 34.207031 33.160156 34.773438 33.683594 34.949219 C 34.207031 35.125 34.773438 34.839844 34.949219 34.316406 C 34.949219 34.316406 35.453125 32.800781 35.960938 31.027344 C 36.46875 29.25 37 27.332031 37 26 C 37 25.949219 36.996094 25.898438 36.988281 25.847656 C 36.960938 24.621094 36.601563 23.265625 35.777344 22.121094 C 34.917969 20.929688 33.480469 20 31.640625 20 C 28.304688 20 26.25 22.339844 26.25 22.339844 C 26.089844 22.519531 26 22.757813 26 23 L 26 26.390625 C 25.894531 26.355469 25.792969 26.3125 25.683594 26.285156 C 25.5 26.105469 25.257813 26.003906 25 26 C 24.742188 26.003906 24.5 26.105469 24.316406 26.285156 C 24.207031 26.3125 24.105469 26.355469 24 26.390625 L 24 23 C 24 22.757813 23.910156 22.519531 23.75 22.339844 C 23.75 22.339844 21.695313 20 18.359375 20 Z M 18.359375 22 C 20.480469 22 21.695313 23.152344 22 23.464844 L 22 27 L 18.011719 27 C 18.007813 27 18.003906 27 18 27 C 17.988281 27 17.980469 27 17.96875 27 C 17.96875 27 17.066406 27.019531 16.230469 26.824219 C 15.808594 26.726563 15.429688 26.566406 15.234375 26.414063 C 15.035156 26.261719 15 26.21875 15 26 C 15 25.144531 15.289063 24.066406 15.847656 23.285156 C 16.40625 22.507813 17.148438 22 18.359375 22 Z M 31.640625 22 C 32.851563 22 33.59375 22.507813 34.152344 23.285156 C 34.710938 24.066406 35 25.144531 35 26 C 35 26.21875 34.964844 26.261719 34.765625 26.414063 C 34.570313 26.566406 34.191406 26.726563 33.769531 26.824219 C 32.933594 27.019531 32.03125 27 32.03125 27 C 32.019531 27 32.011719 27 32 27 L 28 27 L 28 23.464844 C 28.304688 23.152344 29.519531 22 31.640625 22 Z M 25 28 C 25.566406 28 26 28.433594 26 29 L 26 30 L 24 30 L 24 29 C 24 28.433594 24.433594 28 25 28 Z M 23.515625 32 L 24 32 L 24 37 L 23 37 L 23 32.722656 Z M 26 32 L 26.484375 32 L 27 32.722656 L 27 37 L 26 37 Z M 21 35.523438 L 21 37 L 19.945313 37 Z M 29 35.523438 L 30.054688 37 L 29 37 Z"
-                  ></path>
+                    d="M 24 2 C 22.90625 2 22 2.90625 22 4 L 22 4.046875 C 16.664063 4.492188 12.289063 8.46875 11.242188 13.625 C 11.171875 13.902344 11.125 14.203125 11.089844 14.527344 C 11.074219 14.566406 11.023438 14.691406 11.003906 14.917969 C 10.9375 15.726563 10.734375 16.8125 10.484375 17.90625 C 10.40625 18.242188 10.328125 18.578125 10.246094 18.90625 L 10.246094 18.914063 C 10.238281 18.941406 10.226563 18.980469 10.214844 19.035156 C 9.929688 20.148438 9.632813 21.191406 9.40625 21.949219 C 9 22.964844 8.554688 24.039063 8.109375 25.109375 C 7.128906 27.457031 5.871094 30.128906 4.851563 32.589844 C 3.828125 35.042969 3.003906 37.21875 3 38.988281 C 2.996094 39.386719 3.132813 40.554688 4.328125 41.769531 C 5.097656 42.550781 6.273438 43.140625 7.882813 43.71875 C 9.355469 44.253906 11.261719 44.714844 13.371094 45.097656 C 13.652344 45.617188 14.042969 46.070313 14.554688 46.378906 C 15.289063 46.820313 16.1875 47.121094 17.234375 47.355469 C 19.332031 47.820313 22.03125 48 25 48 C 27.964844 48 30.667969 47.832031 32.765625 47.371094 C 33.8125 47.136719 34.714844 46.839844 35.453125 46.394531 C 35.964844 46.082031 36.355469 45.621094 36.636719 45.097656 C 38.742188 44.714844 40.648438 44.253906 42.121094 43.71875 C 43.722656 43.140625 44.902344 42.550781 45.671875 41.769531 C 46.867188 40.554688 47.003906 39.386719 47 38.988281 C 46.996094 37.21875 46.171875 35.046875 45.152344 32.589844 C 44.128906 30.128906 42.871094 27.457031 41.894531 25.109375 C 41.445313 24.035156 41 22.957031 40.589844 21.9375 C 40.574219 21.886719 40.5625 21.839844 40.546875 21.796875 L 40.546875 21.785156 C 40.550781 21.808594 40.546875 21.792969 40.535156 21.757813 L 40.535156 21.753906 C 40.5 21.625 40.457031 21.480469 40.410156 21.328125 C 40.398438 21.285156 40.386719 21.242188 40.375 21.191406 L 40.371094 21.1875 C 40.332031 21.054688 40.292969 20.917969 40.253906 20.78125 L 40.253906 20.777344 C 40.242188 20.734375 40.230469 20.699219 40.222656 20.667969 L 40.222656 20.664063 L 40.21875 20.660156 C 40.171875 20.488281 40.125 20.308594 40.074219 20.121094 L 40.070313 20.109375 C 40.058594 20.0625 40.046875 20.023438 40.039063 19.992188 L 40.035156 19.988281 L 40.035156 19.984375 C 39.996094 19.832031 39.953125 19.675781 39.910156 19.515625 C 39.894531 19.457031 39.882813 19.40625 39.875 19.363281 L 39.871094 19.355469 L 39.871094 19.347656 C 39.816406 19.152344 39.765625 18.949219 39.71875 18.75 L 39.707031 18.714844 L 39.707031 18.707031 C 39.660156 18.523438 39.617188 18.332031 39.570313 18.140625 L 39.566406 18.136719 C 39.554688 18.078125 39.542969 18.023438 39.53125 17.976563 L 39.53125 17.964844 C 39.273438 16.851563 39.066406 15.742188 38.996094 14.917969 C 38.976563 14.6875 38.925781 14.5625 38.910156 14.527344 C 38.875 14.203125 38.828125 13.898438 38.757813 13.625 C 37.710938 8.464844 33.335938 4.492188 28 4.046875 L 28 4 C 28 2.90625 27.09375 2 26 2 Z M 24 4 L 26 4 L 26 18 L 24 18 Z M 22 6.054688 L 22 17.59375 C 21.085938 17.3125 19.738281 17.003906 18 17 L 17 17 C 15.453125 17 13.832031 17.449219 12.4375 18.34375 C 12.695313 17.203125 12.914063 16.066406 12.996094 15.082031 C 12.992188 15.125 12.984375 15.164063 13.027344 15 L 13.046875 14.921875 L 13.058594 14.84375 C 13.089844 14.535156 13.132813 14.285156 13.1875 14.082031 L 13.191406 14.058594 L 13.199219 14.03125 C 14.054688 9.789063 17.621094 6.496094 22 6.054688 Z M 28 6.054688 C 32.378906 6.496094 35.945313 9.789063 36.800781 14.03125 L 36.808594 14.058594 L 36.8125 14.082031 C 36.867188 14.285156 36.910156 14.539063 36.941406 14.84375 L 36.953125 14.921875 L 36.972656 15 C 37.015625 15.164063 37.007813 15.125 37.003906 15.082031 C 37.085938 16.066406 37.304688 17.207031 37.566406 18.347656 C 36.167969 17.449219 34.546875 17 33 17 L 32 17 C 30.261719 17.003906 28.914063 17.3125 28 17.59375 Z M 19.953125 7.996094 C 19.863281 8 19.773438 8.019531 19.6875 8.046875 C 19.6875 8.046875 16.351563 9.136719 15.0625 12.65625 C 14.871094 13.171875 15.136719 13.75 15.65625 13.9375 C 16.171875 14.128906 16.75 13.863281 16.9375 13.34375 C 17.898438 10.738281 20.3125 9.953125 20.3125 9.953125 C 20.777344 9.808594 21.070313 9.347656 21.007813 8.867188 C 20.945313 8.382813 20.542969 8.015625 20.054688 7.996094 C 20.023438 7.996094 19.988281 7.996094 19.953125 7.996094 Z M 17 19 L 18 19 C 20.644531 19 22.574219 19.90625 22.574219 19.90625 L 22.777344 20 L 27.222656 20 L 27.425781 19.90625 C 27.425781 19.90625 29.355469 19 32 19 L 33 19 C 35.226563 19 37.667969 20.167969 38.726563 22.664063 C 39.144531 23.707031 39.597656 24.796875 40.046875 25.878906 C 41.050781 28.292969 42.308594 30.960938 43.304688 33.359375 C 44.300781 35.753906 45 37.945313 45 39 L 45 39.019531 C 45 38.933594 45.058594 39.542969 44.25 40.367188 L 44.246094 40.367188 C 43.90625 40.710938 42.898438 41.3125 41.4375 41.839844 C 40.15625 42.304688 38.503906 42.738281 36.59375 43.097656 C 36.445313 42.878906 36.226563 42.5 35.949219 41.675781 C 35.808594 41.25 35.402344 40.96875 34.953125 40.988281 C 34.632813 41.003906 34.34375 41.164063 34.164063 41.429688 C 33.988281 41.691406 33.945313 42.023438 34.050781 42.324219 C 34.367188 43.242188 34.683594 43.804688 34.925781 44.183594 C 34.871094 44.320313 34.757813 44.472656 34.414063 44.683594 C 33.976563 44.949219 33.253906 45.214844 32.335938 45.417969 C 30.496094 45.820313 27.886719 46 25 46 C 22.117188 46 19.507813 45.8125 17.671875 45.402344 C 16.75 45.199219 16.027344 44.929688 15.589844 44.667969 C 15.339844 44.515625 15.214844 44.394531 15.132813 44.289063 C 15.667969 43.742188 16.480469 42.761719 16.953125 41.308594 C 17.054688 41 17.003906 40.664063 16.816406 40.402344 C 16.628906 40.140625 16.324219 39.984375 16 39.984375 C 15.5625 39.988281 15.175781 40.273438 15.046875 40.691406 C 14.597656 42.09375 13.742188 42.902344 13.496094 43.113281 C 11.546875 42.753906 9.863281 42.308594 8.5625 41.839844 C 7.101563 41.3125 6.089844 40.710938 5.75 40.367188 C 4.941406 39.542969 5 38.933594 5 39.019531 L 5 39 C 5 37.945313 5.699219 35.753906 6.695313 33.359375 C 7.691406 30.960938 8.949219 28.292969 9.953125 25.878906 C 10.410156 24.792969 10.863281 23.695313 11.28125 22.644531 C 12.347656 20.164063 14.777344 19 17 19 Z M 18.359375 20 C 16.519531 20 15.082031 20.929688 14.222656 22.121094 C 13.402344 23.257813 13.042969 24.605469 13.011719 25.828125 C 13.011719 25.832031 13.011719 25.832031 13.011719 25.832031 C 13.003906 25.886719 13 25.945313 13 26 C 13 27.332031 13.53125 29.25 14.039063 31.027344 C 14.546875 32.800781 15.050781 34.316406 15.050781 34.316406 C 15.226563 34.839844 15.792969 35.125 16.316406 34.949219 C 16.839844 34.773438 17.125 34.207031 16.949219 33.683594 C 16.949219 33.683594 16.453125 32.199219 15.960938 30.472656 C 15.792969 29.890625 15.652344 29.285156 15.511719 28.6875 C 15.597656 28.710938 15.6875 28.75 15.769531 28.769531 C 16.851563 29.023438 17.785156 29.003906 17.929688 29 C 17.953125 29 17.976563 29 18 29 C 18 29 18.03125 29 18.03125 29 L 22 29 L 22 30.679688 L 17.1875 37.417969 C 16.96875 37.722656 16.9375 38.125 17.109375 38.457031 C 17.28125 38.792969 17.625 39 18 39 L 32 39 C 32.375 39 32.71875 38.792969 32.890625 38.457031 C 33.0625 38.125 33.03125 37.722656 32.8125 37.417969 L 28 30.679688 L 28 29 L 32 29 C 32 29 32.027344 28.996094 32.03125 29 C 32.125 29 33.117188 29.03125 34.230469 28.769531 C 34.3125 28.75 34.402344 28.710938 34.488281 28.6875 C 34.347656 29.285156 34.207031 29.890625 34.039063 30.472656 C 33.546875 32.199219 33.050781 33.683594 33.050781 33.683594 C 32.875 34.207031 33.160156 34.773438 33.683594 34.949219 C 34.207031 35.125 34.773438 34.839844 34.949219 34.316406 C 34.949219 34.316406 35.453125 32.800781 35.960938 31.027344 C 36.46875 29.25 37 27.332031 37 26 C 37 25.949219 36.996094 25.898438 36.988281 25.847656 C 36.960938 24.621094 36.601563 23.265625 35.777344 22.121094 C 34.917969 20.929688 33.480469 20 31.640625 20 C 28.304688 20 26.25 22.339844 26.25 22.339844 C 26.089844 22.519531 26 22.757813 26 23 L 26 26.390625 C 25.894531 26.355469 25.792969 26.3125 25.683594 26.285156 C 25.5 26.105469 25.257813 26.003906 25 26 C 24.742188 26.003906 24.5 26.105469 24.316406 26.285156 C 24.207031 26.3125 24.105469 26.355469 24 26.390625 L 24 23 C 24 22.757813 23.910156 22.519531 23.75 22.339844 C 23.75 22.339844 21.695313 20 18.359375 20 Z M 18.359375 22 C 20.480469 22 21.695313 23.152344 22 23.464844 L 22 27 L 18.011719 27 C 18.007813 27 18.003906 27 18 27 C 17.988281 27 17.980469 27 17.96875 27 C 17.96875 27 17.066406 27.019531 16.230469 26.824219 C 15.808594 26.726563 15.429688 26.566406 15.234375 26.414063 C 15.035156 26.261719 15 26.21875 15 26 C 15 25.144531 15.289063 24.066406 15.847656 23.285156 C 16.40625 22.507813 17.148438 22 18.359375 22 Z M 31.640625 22 C 32.851563 22 33.59375 22.507813 34.152344 23.285156 C 34.710938 24.066406 35 25.144531 35 26 C 35 26.21875 34.964844 26.261719 34.765625 26.414063 C 34.570313 26.566406 34.191406 26.726563 33.769531 26.824219 C 32.933594 27.019531 32.03125 27 32.03125 27 C 32.019531 27 32.011719 27 32 27 L 28 27 L 28 23.464844 C 28.304688 23.152344 29.519531 22 31.640625 22 Z M 25 28 C 25.566406 28 26 28.433594 26 29 L 26 30 L 24 30 L 24 29 C 24 28.433594 24.433594 28 25 28 Z M 23.515625 32 L 24 32 L 24 37 L 23 37 L 23 32.722656 Z M 26 32 L 26.484375 32 L 27 32.722656 L 27 37 L 26 37 Z M 21 35.523438 L 21 37 L 19.945313 37 Z M 29 35.523438 L 30.054688 37 L 29 37 Z">
+                  </path>
                 </svg>
               </span>
             </div>
@@ -36,10 +23,8 @@
                 sagittis viverra duis.
               </p>
               <div class="mt-6">
-                <router-link
-                  to="/people/1"
-                  class="inline-flex px-4 py-2 border border-transparent text-base font-medium rounded-md shadow-sm bg-red-700 hover:bg-red-800"
-                >
+                <router-link to="/people/1"
+                  class="inline-flex px-4 py-2 border border-transparent text-base font-medium rounded-md shadow-sm bg-red-700 hover:bg-red-800">
                   See more
                 </router-link>
               </div>
@@ -47,30 +32,20 @@
           </div>
         </div>
         <div class="mt-12 sm:mt-16 lg:mt-0">
-          <div
-            class="pl-4 -mr-48 sm:pl-6 md:-mr-16 lg:px-0 lg:m-0 lg:relative lg:h-full"
-          >
+          <div class="pl-4 -mr-48 sm:pl-6 md:-mr-16 lg:px-0 lg:m-0 lg:relative lg:h-full">
             <img
               class="w-full object-cover rounded-xl shadow-xl ring-1 ring-black ring-opacity-5 lg:absolute lg:left-0 lg:h-full lg:w-auto lg:max-w-none"
-              :src="characterUrl"
-              alt="Image of dark vador"
-            />
+              :src="characterUrl" alt="Image of dark vador" />
           </div>
         </div>
       </div>
     </div>
     <div class="mt-24">
-      <div
-        class="lg:mx-auto lg:max-w-7xl lg:px-8 lg:grid lg:grid-cols-2 lg:grid-flow-col-dense lg:gap-24"
-      >
-        <div
-          class="px-4 max-w-xl mx-auto sm:px-6 lg:py-32 lg:max-w-none lg:mx-0 lg:px-0 lg:col-start-2"
-        >
+      <div class="lg:mx-auto lg:max-w-7xl lg:px-8 lg:grid lg:grid-cols-2 lg:grid-flow-col-dense lg:gap-24">
+        <div class="px-4 max-w-xl mx-auto sm:px-6 lg:py-32 lg:max-w-none lg:mx-0 lg:px-0 lg:col-start-2">
           <div>
             <div class="inline-flex justify-center">
-              <span
-                class="h-12 w-12 rounded-md flex items-center justify-center bg-red-700 p-2"
-              >
+              <span class="h-12 w-12 rounded-md flex items-center justify-center bg-red-700 p-2">
                 <img :src="vehicleIconUrl" class="h-full w-full object-cover" />
               </span>
             </div>
@@ -83,10 +58,8 @@
                 at.
               </p>
               <div class="mt-6">
-                <a
-                  href="#"
-                  class="inline-flex px-4 py-2 border border-transparent text-base font-medium rounded-md shadow-sm bg-red-700 hover:bg-red-800"
-                >
+                <a href="#"
+                  class="inline-flex px-4 py-2 border border-transparent text-base font-medium rounded-md shadow-sm bg-red-700 hover:bg-red-800">
                   See more
                 </a>
               </div>
@@ -94,32 +67,22 @@
           </div>
         </div>
         <div class="mt-12 sm:mt-16 lg:mt-0 lg:col-start-1">
-          <div
-            class="pr-4 -ml-48 sm:pr-6 md:-ml-16 lg:px-0 lg:m-0 lg:relative lg:h-full"
-          >
+          <div class="pr-4 -ml-48 sm:pr-6 md:-ml-16 lg:px-0 lg:m-0 lg:relative lg:h-full">
             <img
               class="w-full rounded-xl shadow-xl ring-1 ring-black ring-opacity-5 lg:absolute lg:right-0 lg:h-full lg:w-auto lg:max-w-none"
-              :src="vehicleUrl"
-              alt="Customer profile user interface"
-            />
+              :src="vehicleUrl" alt="Customer profile user interface" />
           </div>
         </div>
       </div>
     </div>
 
     <div class="mt-24">
-      <div
-        class="lg:mx-auto lg:max-w-7xl lg:px-8 lg:grid lg:grid-cols-2 lg:grid-flow-col-dense lg:gap-24"
-      >
-        <div
-          class="px-4 max-w-xl mx-auto sm:px-6 lg:py-16 lg:max-w-none lg:mx-0 lg:px-0"
-        >
+      <div class="lg:mx-auto lg:max-w-7xl lg:px-8 lg:grid lg:grid-cols-2 lg:grid-flow-col-dense lg:gap-24">
+        <div class="px-4 max-w-xl mx-auto sm:px-6 lg:py-16 lg:max-w-none lg:mx-0 lg:px-0">
           <div>
             <div class="flex justify-center">
-              <span
-                class="h-12 w-12 rounded-md flex items-center justify-center bg-red-700 p-1"
-              >
-                <FilmIcon></FilmIcon>
+              <span class="h-12 w-12 rounded-md flex items-center justify-center bg-red-700 p-1">
+                <Icon name="heroicons:film" />
               </span>
             </div>
             <div class="mt-6">
@@ -130,10 +93,8 @@
                 sagittis viverra duis.
               </p>
               <div class="mt-6">
-                <router-link
-                  to="/people/1"
-                  class="inline-flex px-4 py-2 border border-transparent text-base font-medium rounded-md shadow-sm bg-red-700 hover:bg-red-800"
-                >
+                <router-link to="/people/1"
+                  class="inline-flex px-4 py-2 border border-transparent text-base font-medium rounded-md shadow-sm bg-red-700 hover:bg-red-800">
                   See more
                 </router-link>
               </div>
@@ -141,14 +102,10 @@
           </div>
         </div>
         <div class="mt-12 sm:mt-16 lg:mt-0">
-          <div
-            class="pl-4 -mr-48 sm:pl-6 md:-mr-16 lg:px-0 lg:m-0 lg:relative lg:h-full"
-          >
+          <div class="pl-4 -mr-48 sm:pl-6 md:-mr-16 lg:px-0 lg:m-0 lg:relative lg:h-full">
             <img
               class="w-full object-cover rounded-xl shadow-xl ring-1 ring-black ring-opacity-5 lg:absolute lg:left-0 lg:h-full lg:w-auto lg:max-w-none"
-              :src="movieUrl"
-              alt="Image of dark vador"
-            />
+              :src="movieUrl" alt="Image of dark vador" />
           </div>
         </div>
       </div>
@@ -157,7 +114,7 @@
 </template>
 
 <script setup>
-import { FilmIcon } from "@heroicons/vue/outline";
+import { Icon } from '@iconify/vue';
 import vehicleUrl from "../../assets/img/vehicles/14.png";
 import movieUrl from "../../assets/img/films/2.png";
 import characterUrl from "../../assets/img/characters/4.png";
