@@ -1,4 +1,4 @@
-import axios, { AxiosInstance } from "axios";
+import axios from "axios";
 const apiClient = axios.create({
   baseURL: "https://swapi.dev/api",
   headers: {
